refactor(website): text input
